Initial commit - basic dev environment
Starting the project with a basic development environment that I use
for experimenting with new tools. Running `make` will generate a dev
container that has some useful tools built into it, and then drop me
into a bash shell there.  The local directory will be mapped into it
along with my aws credentials.

The makefile also contains a simple target to run a terraform plan, but
does not yet include anything to do with environments or workspaces
because this really is the most basic thing I needed.