Add bootstrap steps for IAM
Updated the iam policy so that the bucket name can be provided to it
and updated the bootstrap script to fill in the details.  The script
now also creates the user and an access key.

At this point, once the file has been run, I have enough pieces to
have a backend setup.