Hide bucket name in the backend
The backend file is now a template, which I'll fill in with an
environment variable. This allows me to hide the name of my state bucket
from anyone with access to the git repo.

Perhaps this isn't necessary, but it does give me a chance to show-off
my Makefile skills :-)

In addition, I've added a dependency on a hard-coded workspace.  I'm going
to make this a parameter very soon.