Add flag for uploading website content
Allowing Terraform to upload the content to the s3 bucket, so that the
static site can be uploaded and managed from one TF command is an easy
thing to setup, but is not a great idea if there are teams collaborating
on the same content.  A branch on which an engineer is tweaking some
networking parameters, and another branch in which a content-writer is
updating the HTML, would always be in conflict with each other.

However, for the purposes of this homework exercise, and with just one
person working on the repo (me) it's probably ok.  I've added a flag so
that I can turn this off and implement an alternative outside of Terraform
later.