Add a Cloudfront Distribution
Adds a CF Distribution for the S3 bucket.  This allows the bucket to continue
being not-public access, and provides POP caching.

I had quite a lot of trouble getting this to work -- there was a problem, somewhere,
in the Cloudfront configuration and that caused Cloudfront to return a 503(!) error.
Terraform retried, over and over.  There's been quite a lot of trial and error to get
here - found it difficult to surface the actual error.