Use my personal domain as an alias
Created an alias for the CF distribution, and added a wildcard cert
to ACM in us-east-1

The domain "grahamandsarah.com" is one I've owned for a long time
and was originally used for our wedding planning!

Rather than creating a domain through Route53 & TF, I took the shortcut
of reusing one I already own. Note that to get to this point, I had to
manually create the ACM cert, and then manually setup the CNAME to prove
domain ownership.