Add nginx container for serving the static site
So that the static site can be served from a host, rather than a CDN,
build a docker container that holds the HTML files.  Add an ECR to hold
the container, and add Make targets to build and then push the image.

The mechanism here is to _not_ use Terraform for the Docker build and
push steps - I like to keep separation and use TF for Infrastructure and
Make for building.